+++ backend/integration_tests/src/lib.rs
@@ -32,6 +32,7 @@ mod set_message_reminder_tests;
 mod setup;
 mod storage;
 mod suspend_user_tests;
+mod tip_message_tests;
 mod update_group_tests;
 mod update_profile_tests;
 mod utils;